+++ internal/certificate/resource_test.go
@@ -2,7 +2,6 @@ package certificate_test
 
 import (
 	"fmt"
-	"os"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
@@ -14,12 +13,6 @@ import (
 	"github.com/hetznercloud/terraform-provider-hcloud/internal/testmux"
 	"github.com/hetznercloud/terraform-provider-hcloud/internal/testsupport"
 	"github.com/hetznercloud/terraform-provider-hcloud/internal/testtemplate"
-)
-
-var (
-	// Tests using this value should skip if empty.
-	// The domain specified here must be available in Hetzner DNS of the account running the tests.
-	certDomain = os.Getenv("CERT_DOMAIN")
 )
 
 func TestAccCertificateResource_Uploaded(t *testing.T) {
@@ -115,14 +108,14 @@ func TestAccCertificateResource_Uploaded_ChangeCertRequiresNewResource(t *testin
 }
 
 func TestAccCertificateResource_Managed(t *testing.T) {
-	if certDomain == "" {
-		t.Skip("Skipping because CERT_DOMAIN is not set")
+	if teste2e.TestCertificateDomain == "" {
+		t.Skip("Skipping because TEST_CERTIFICATE_DOMAIN is not set")
 	}
 
 	var cert hcloud.Certificate
 
 	res := certificate.NewManagedRData(t, "basic-managed-cert", []string{
-		fmt.Sprintf("tftest-%d.%s", acctest.RandInt(), certDomain),
+		fmt.Sprintf("tftest-%d.%s", acctest.RandInt(), teste2e.TestCertificateDomain),
 	})
 	resRenamed := &certificate.RDataManaged{
 		Name:        res.Name + "-renamed",